The task-card dispatcher moves in beside its nine leaves
`TaskCard.tsx` was marooned at `components/` root while every skin it
dispatches sat in `components/cards/`. Both halves are now
`components/taskCard/`. Pure move -- no rendered output changed.

`.design-sync/config.json` and `.ds-kit/index.tsx` follow for the five
entries whose paths resolve case-exactly. `SNIDETaskCard` and `UATaskCard`
are left alone: both were already dangling on `origin/main` (the files are
`SnideTaskCard.tsx` / `UaTaskCard.tsx`), so they belong to the 39-entry
stale backlog, not to this move.
