+++ .design-sync/config.json
@@ -16,13 +16,13 @@
     "CommentThread": "src/components/comments/CommentThread.tsx",
     "CredentialCard": "src/components/CredentialCard.tsx",
     "DefaultSigil": "src/components/sigil/DefaultSigil.tsx",
-    "DefaultTaskCard": "src/components/cards/DefaultTaskCard.tsx",
+    "DefaultTaskCard": "src/components/taskCard/DefaultTaskCard.tsx",
     "EphemeristsAvatar": "src/components/avatar/EphemeristsAvatar.tsx",
     "EphemeristsBackdrop": "src/components/backdrop/EphemeristsBackdrop.tsx",
     "EphemeristsComment": "src/components/comments/voices/EphemeristsComment.tsx",
     "EphemeristsFactionHero": "src/components/cards/EphemeristsFactionHero.tsx",
     "EphemeristsFeedFrame": "src/components/feed/EphemeristsFeedFrame.tsx",
-    "EphemeristsTaskCard": "src/components/cards/EphemeristsTaskCard.tsx",
+    "EphemeristsTaskCard": "src/components/taskCard/EphemeristsTaskCard.tsx",
     "EphemeristsVote": "src/components/vote/EphemeristsVote.tsx",
     "EverymenAvatar": "src/components/avatar/EverymenAvatar.tsx",
     "EverymenBackdrop": "src/components/backdrop/EverymenBackdrop.tsx",
@@ -30,7 +30,7 @@
     "EverymenComment": "src/components/comments/voices/EverymenComment.tsx",
     "EverymenFactionHero": "src/components/cards/EverymenFactionHero.tsx",
     "EverymenFeedFrame": "src/components/feed/EverymenFeedFrame.tsx",
-    "EverymenTaskCard": "src/components/cards/EverymenTaskCard.tsx",
+    "EverymenTaskCard": "src/components/taskCard/EverymenTaskCard.tsx",
     "EverymenVote": "src/components/vote/EverymenVote.tsx",
     "FactionAvatar": "src/components/avatar/FactionAvatar.tsx",
     "FactionBackdrop": "src/components/backdrop/FactionBackdrop.tsx",
@@ -59,7 +59,7 @@
     "SingularityComment": "src/components/comments/voices/SingularityComment.tsx",
     "SingularityFactionHero": "src/components/cards/SingularityFactionHero.tsx",
     "SingularityFeedFrame": "src/components/feed/SingularityFeedFrame.tsx",
-    "SingularityTaskCard": "src/components/cards/SingularityTaskCard.tsx",
+    "SingularityTaskCard": "src/components/taskCard/SingularityTaskCard.tsx",
     "SingularityVote": "src/components/vote/SingularityVote.tsx",
     "SnideAvatar": "src/components/avatar/SnideAvatar.tsx",
     "SnideBackdrop": "src/components/backdrop/SnideBackdrop.tsx",
@@ -69,7 +69,7 @@
     "SnideMasthead": "src/components/cards/SnideMasthead.tsx",
     "SNIDETaskCard": "src/components/cards/SNIDETaskCard.tsx",
     "SnideVote": "src/components/vote/SnideVote.tsx",
-    "TaskCard": "src/components/TaskCard.tsx",
+    "TaskCard": "src/components/taskCard/TaskCard.tsx",
     "TaskCrown": "src/components/cards/TaskCrown.tsx",
     "UAAvatar": "src/components/avatar/UAAvatar.tsx",
     "UABackdrop": "src/components/backdrop/UABackdrop.tsx",
@@ -86,7 +86,7 @@
     "WowComment": "src/components/comments/voices/WowComment.tsx",
     "WowFactionHero": "src/components/cards/WowFactionHero.tsx",
     "WowFeedFrame": "src/components/feed/WowFeedFrame.tsx",
-    "WowTaskCard": "src/components/cards/WowTaskCard.tsx",
+    "WowTaskCard": "src/components/taskCard/WowTaskCard.tsx",
     "WowVote": "src/components/vote/WowVote.tsx",
     "DefaultCreateCharacter": "src/pages/characterPaths/mobileArchetypes/DefaultCreateCharacter.tsx",
     "DefaultEditCharacter": "src/pages/characterPaths/mobileArchetypes/DefaultEditCharacter.tsx",
