+++ frontend/.ds-kit/index.tsx
@@ -7,13 +7,13 @@ export { default as CollaborationCard } from "../src/components/CollaborationCar
 export { default as CommentThread } from "../src/components/comments/CommentThread";
 export { default as CredentialCard } from "../src/components/CredentialCard";
 export { default as DefaultSigil } from "../src/components/sigil/DefaultSigil";
-export { default as DefaultTaskCard } from "../src/components/cards/DefaultTaskCard";
+export { default as DefaultTaskCard } from "../src/components/taskCard/DefaultTaskCard";
 export { default as EphemeristsAvatar } from "../src/components/avatar/EphemeristsAvatar";
 export { default as EphemeristsBackdrop } from "../src/components/backdrop/EphemeristsBackdrop";
 export { default as EphemeristsComment } from "../src/components/comments/voices/EphemeristsComment";
 export { default as EphemeristsFactionHero } from "../src/components/cards/EphemeristsFactionHero";
 export { default as EphemeristsFeedFrame } from "../src/components/feed/EphemeristsFeedFrame";
-export { default as EphemeristsTaskCard } from "../src/components/cards/EphemeristsTaskCard";
+export { default as EphemeristsTaskCard } from "../src/components/taskCard/EphemeristsTaskCard";
 export { default as EphemeristsVote } from "../src/components/vote/EphemeristsVote";
 export { default as EverymenAvatar } from "../src/components/avatar/EverymenAvatar";
 export { default as EverymenBackdrop } from "../src/components/backdrop/EverymenBackdrop";
@@ -21,7 +21,7 @@ export { default as EverymenCard } from "../src/components/cards/EverymenFaction
 export { default as EverymenComment } from "../src/components/comments/voices/EverymenComment";
 export { default as EverymenFactionHero } from "../src/components/cards/EverymenFactionHero";
 export { default as EverymenFeedFrame } from "../src/components/feed/EverymenFeedFrame";
-export { default as EverymenTaskCard } from "../src/components/cards/EverymenTaskCard";
+export { default as EverymenTaskCard } from "../src/components/taskCard/EverymenTaskCard";
 export { default as EverymenVote } from "../src/components/vote/EverymenVote";
 export { default as FactionAvatar } from "../src/components/avatar/FactionAvatar";
 export { default as FactionBackdrop } from "../src/components/backdrop/FactionBackdrop";
@@ -50,7 +50,7 @@ export { default as SingularityBackdrop } from "../src/components/backdrop/Singu
 export { default as SingularityComment } from "../src/components/comments/voices/SingularityComment";
 export { default as SingularityFactionHero } from "../src/components/cards/SingularityFactionHero";
 export { default as SingularityFeedFrame } from "../src/components/feed/SingularityFeedFrame";
-export { default as SingularityTaskCard } from "../src/components/cards/SingularityTaskCard";
+export { default as SingularityTaskCard } from "../src/components/taskCard/SingularityTaskCard";
 export { default as SingularityVote } from "../src/components/vote/SingularityVote";
 export { default as SnideAvatar } from "../src/components/avatar/SnideAvatar";
 export { default as SnideBackdrop } from "../src/components/backdrop/SnideBackdrop";
@@ -60,7 +60,7 @@ export { default as SnideFeedFrame } from "../src/components/feed/SnideFeedFrame
 export { default as SnideMasthead } from "../src/components/cards/SnideMasthead";
 export { default as SNIDETaskCard } from "../src/components/cards/SNIDETaskCard";
 export { default as SnideVote } from "../src/components/vote/SnideVote";
-export { default as TaskCard } from "../src/components/TaskCard";
+export { default as TaskCard } from "../src/components/taskCard/TaskCard";
 export { TaskCrown } from "../src/components/cards/TaskCrown";
 export { default as UAAvatar } from "../src/components/avatar/UAAvatar";
 export { default as UABackdrop } from "../src/components/backdrop/UABackdrop";
@@ -77,7 +77,7 @@ export { default as WowBackdrop } from "../src/components/backdrop/WowBackdrop";
 export { default as WowComment } from "../src/components/comments/voices/WowComment";
 export { default as WowFactionHero } from "../src/components/cards/WowFactionHero";
 export { default as WowFeedFrame } from "../src/components/feed/WowFeedFrame";
-export { default as WowTaskCard } from "../src/components/cards/WowTaskCard";
+export { default as WowTaskCard } from "../src/components/taskCard/WowTaskCard";
 export { default as WowVote } from "../src/components/vote/WowVote";
 export { default as DefaultCreateCharacter } from "../src/pages/characterPaths/mobileArchetypes/DefaultCreateCharacter";
 export { default as DefaultEditCharacter } from "../src/pages/characterPaths/mobileArchetypes/DefaultEditCharacter";
