+++ frontend/src/components/taskCard/EverymenTaskCard.tsx
@@ -1,6 +1,6 @@
 import type { CSSProperties } from "react";
 import { Link } from "react-router-dom";
-import type { CardProps } from "../TaskCard";
+import type { CardProps } from "./TaskCard";
 import i18n from "../../i18n";
 import { isNeutralMultiplier } from "../../utils/points";
 import { useFormFactor } from "../../hooks/useFormFactor";
