+++ frontend/src/components/taskCard/TaskCard.tsx
@@ -1,10 +1,10 @@
-import type { TaskOut } from '../api/tasks'
-import { useAuth } from '../auth/AuthContext'
-import { useAdminMode } from '../auth/AdminModeContext'
-import DefaultTaskCard from './cards/DefaultTaskCard'
-import { factionCssVar, factionFill, factionName } from '../utils/factions'
-import { pickVariant } from '../utils/factionDispatch'
-import { surfaceMap } from '../factions'
+import type { TaskOut } from '../../api/tasks'
+import { useAuth } from '../../auth/AuthContext'
+import { useAdminMode } from '../../auth/AdminModeContext'
+import DefaultTaskCard from './DefaultTaskCard'
+import { factionCssVar, factionFill, factionName } from '../../utils/factions'
+import { pickVariant } from '../../utils/factionDispatch'
+import { surfaceMap } from '../../factions'
 import type { } from 'react'
 
 /**
@@ -59,7 +59,7 @@ export default function TaskCard({
   // renders for every visitor, and a static import put the admin chunk in every
   // logged-out `/tasks` waterfall for a control only a moderator can reach.
   const handleStatusChange = async (newStatus: string) => {
-    const { updateTaskStatus } = await import('../api/admin')
+    const { updateTaskStatus } = await import('../../api/admin')
     await updateTaskStatus(task.id, newStatus)
     window.location.reload()
   }
