+++ frontend/src/components/taskCard/UaTaskCard.tsx
@@ -1,11 +1,11 @@
 import { Link } from "react-router-dom";
-import type { CardProps } from "../TaskCard";
+import type { CardProps } from "./TaskCard";
 import i18n from "../../i18n";
 import { isNeutralMultiplier } from "../../utils/points";
 import { useFormFactor } from "../../hooks/useFormFactor";
 import { Lotus } from "../factionMarks";
 import { UaSigil } from "../sigil/UaSigil";
-import { UA_DISPLAY, UA_EYEBROW, UA_TEXT, UaEnsoScore } from "./uaAtoms";
+import { UA_DISPLAY, UA_EYEBROW, UA_TEXT, UaEnsoScore } from "../cards/uaAtoms";
 
 /**
  * UA — THE VELLUM LEAF (task card v2, #1023).
