+++ frontend/src/factions/albescent.ts
@@ -33,7 +33,7 @@ import type { FactionManifest } from './manifest'
 import { lazyArchetype } from './lazyArchetype'
 
 const AlbescentSelectCard = lazyArchetype(() => import('../components/cards/FactionSelectCard').then((m) => ({ default: m.AlbescentSelectCard })))
-const AlbescentTaskCard = lazyArchetype(() => import('../components/cards/AlbescentTaskCard'))
+const AlbescentTaskCard = lazyArchetype(() => import('../components/taskCard/AlbescentTaskCard'))
 const AlbescentVote = lazyArchetype(() => import('../components/vote/AlbescentVote'))
 const AlbescentPraxisCard = lazyArchetype(() => import('../components/praxisCard/desktop/AlbescentPraxisCard'))
 const AlbescentSeal = lazyArchetype(() => import('../components/metataskSeal/skins/AlbescentSeal'))
