+++ frontend/src/factions/ua.ts
@@ -24,7 +24,7 @@ const UaFieldDesk = lazyArchetype(() => import('../pages/fieldDesk/mobileArchety
 const UaProfileBody = lazyArchetype(() => import('../pages/characterProfile/archetypes/UaProfileBody'))
 const UaScoreStamp = lazyArchetype(() => import('../components/praxisCard/scoreStamp/UaScoreStamp'))
 const UaSeal = lazyArchetype(() => import('../components/metataskSeal/skins/UaSeal'))
-const UaTaskCard = lazyArchetype(() => import('../components/cards/UaTaskCard'))
+const UaTaskCard = lazyArchetype(() => import('../components/taskCard/UaTaskCard'))
 const UaPraxisDetail = lazyArchetype(() => import('../pages/praxisDetail/archetypes/UaPraxisDetail'))
 const UaTaskDetail = lazyArchetype(() => import('../pages/taskDetail/archetypes/UaTaskDetail'))
 const UaVote = lazyArchetype(() => import('../components/vote/UaVote'))
