+++ frontend/src/pages/Home.tsx
@@ -9,7 +9,7 @@ import { loginWithGoogle, devLogin } from '../api/auth'
 import { computeFactionMultiplier } from '../utils/points'
 import { extractError } from '../utils/errors'
 import PraxisCard from '../components/PraxisCard'
-import TaskCard from '../components/TaskCard'
+import TaskCard from '../components/taskCard/TaskCard'
 import ActivityTicker from '../components/home/ActivityTicker'
 
 /** "Featured Praxis" / "Newest Task" header — marker title + rule + see-all link. */
