+++ frontend/src/pages/characterProfile/archetypes/DefaultProfileBody.tsx
@@ -39,7 +39,7 @@ import type { BadgeOut } from '../../../api/auth'
 import { badgeArtFor } from '../../../components/badges/badgeArt'
 import CredentialCard from '../../../components/CredentialCard'
 import PraxisCard from '../../../components/PraxisCard'
-import TaskCard from '../../../components/TaskCard'
+import TaskCard from '../../../components/taskCard/TaskCard'
 import { useFormFactor } from '../../../hooks/useFormFactor'
 import { factionFill, factionName, isKnownFaction } from '../../../utils/factions'
 import { mediaUrl } from '../../../utils/media'
