+++ frontend/src/pages/factionDetail/archetypes/CovenFactionBody.tsx
@@ -1,7 +1,7 @@
 import { useState, type CSSProperties, type ReactNode } from "react";
 import { Trans, useTranslation } from "react-i18next";
 import { Link } from "react-router-dom";
-import TaskCard from "../../../components/TaskCard";
+import TaskCard from "../../../components/taskCard/TaskCard";
 import PraxisCard from "../../../components/PraxisCard";
 import { TaskCrown } from "../../../components/cards/TaskCrown";
 import { CovenSigil } from "../../../components/sigil/CovenSigil";
